* docs/INTERNALS: documented splitting of dumps across multiple   holding disks. In particular: * server-src/conffile.c: make sure that disksize and chunksize   are multiple of TAPE_BLOCK_SIZE. * server-src/diskfile.c (find_disk): new function to see if a   given disk is contained in a list structure. * server-src/diskfile.h (find_disk): prototype. * server-src/driverio.h (RQ_MORE_DISK): Added message. * server-src/driverio.h (free_assignedhd): prototype. * server-src/driverio.h (assignedhd_t): new data structure. * server-src/driverio.c (cmdstr): added comamnd RQ-MORE-DISK. * server-src/driverio.c (dumper_cmd): Modified format of CONTINUE * server-src/driverio.c (free_assignedhd): new function to free an   array of assignedhd_t. * server-src/driver.c (roomq): new queue to replace stoppedq. * server-src/driver.c (find_diskspace, assign_holdingdisk,   adjust_diskspace, delete_diskspace): rewrite. * server-src/driver.c (continue_dumps): new function to determine   which dump should be continued. * server-src/driver.c (handle_taper_result): call continue_dumps   to restart a roomq dump. * server-src/driver.c (handle_dumper_result): some rewrite,   modified semantics of NO_ROOM, new message RQ-MORE-DISK. * server-src/dumper.c (main): Modified parsing of FILE_DUMP,   handling of use. * server-src/dumper.c (write_dataptr): NO-ROOM is informational only,   trucate the file to a multiple of TAPE_BLOCK_BYTES. * server-src/dumper.c (update_dataptr): rewrite to emit RQ-MORE-DISK   command.

+++ server-src/conffile.h
@@ -25,7 +25,7 @@
  *			   University of Maryland at College Park
  */
 /*
- * $Id: conffile.h,v 1.24.2.4 1999/06/15 08:08:52 oliva Exp $
+ * $Id: conffile.h,v 1.24.2.5 1999/08/21 20:40:32 martinea Exp $
  *
  * interface for config file reading code
  */
@@ -165,8 +165,8 @@ typedef struct holdingdisk_s {
 
     char *comment;
     char *diskdir;
-    int disksize;
-    int chunksize;
+    long disksize;
+    long chunksize;
 
     int s_comment;
     int s_disk;
